Some polish
